core/requirements.txt
--extra-index-url https://${FURY_AUTH_TOKEN}:@deps.kaleido.ai/pypi/
kaleido-core==1.3.10
shadowgen-core==1.3.1
torch==1.7.1
torchvision==0.8.2
numpy>=1.17
opencv-python>=4.1
wandb>=0.10.18
pika==1.0.1
msgpack==0.6.1
pandas>=0.24.2
nibabel==2.4.0
pillow>=7.2.0
